requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docstring_parser==0.16
Flask==3.1.0
Flask-Cors==5.0.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-cloud-vision==3.10.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.26.0  # More compatible with modern builds
packaging==24.2
pdf2image==1.17.0
pillow==11.3.0
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.23.0
pydantic==2.10.3
pydantic_core==2.27.1
PyMuPDF==1.24.14
pyparsing==3.2.0
PyPDF2==3.0.1
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0
python-dotenv==1.1.1
pyttsx3==2.98
pytweening==1.2.0
redis==5.2.0
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
spotipy==2.24.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
vertexai==1.71.1
Werkzeug==3.1.3
Wikipedia-API==0.7.1